Change demo db notification to a blinking red notice in header
The earlier method of notifying the user that he is using hte
demo database was to use differnt header background images. That
was not clear as even I could not rememebr which scheme was for
which.

With this change, if the user is using a demo database a string
"Demo Database" will keep blinking in the header under the title.
IF the user switches to the production database, this message
will not appear.