Use visit charge based on both doc and dept visit status
As per feedback, a visit is to be chared as new visit if a
patient is seeing the same doc in the same dept for the first time.
If the patient has seen same doc in a diff dept, or visited another
doc in the same dept that is not enough to charge as old patient.